related to #48, just preparing abstraction
