+++ bauble/plugins/garden/plant.py
@@ -32,7 +32,6 @@ logger = logging.getLogger(__name__)
 #logger.setLevel(logging.DEBUG)
 
 import gtk
-import glib
 
 from bauble.i18n import _
 from sqlalchemy import and_, func
@@ -370,7 +369,7 @@ acc_type_values = {u'Plant': _('Plant'),
                    None: ''}
 
 
-class Plant(db.Base, db.Serializable):
+class Plant(db.Base, db.Serializable, db.DefiningPictures):
     """
     :Table name: plant
 
@@ -439,43 +438,6 @@ class Plant(db.Base, db.Serializable):
                             backref=backref('plant', uselist=False))
 
     _delimiter = None
-
-    @property
-    def pictures(self):
-        '''a list of gtk.Image objects
-        '''
-
-        pfolder = prefs.prefs[prefs.picture_root_pref]
-        result = []
-        for n in self.notes:
-            if n.category != '<picture>':
-                continue
-            filename = os.path.join(pfolder, n.note)
-            im = gtk.Image()
-            try:
-                pixbuf = gtk.gdk.pixbuf_new_from_file(
-                    os.path.join(prefs.prefs[prefs.picture_root_pref],
-                                 filename))
-                scale_x = pixbuf.get_width() / 400
-                scale_y = pixbuf.get_height() / 400
-                scale = max(scale_x, scale_y, 1)
-                x = int(pixbuf.get_width() / scale)
-                y = int(pixbuf.get_height() / scale)
-                scaled_buf = pixbuf.scale_simple(x, y, gtk.gdk.INTERP_BILINEAR)
-                im.set_from_pixbuf(scaled_buf)
-            except glib.GError, e:
-                logger.debug("picture %s caused glib.GError %s" %
-                             (filename, e))
-                label = _('picture file %s not found.') % filename
-                im = gtk.Label()
-                im.set_text(label)
-            except Exception, e:
-                logger.warning("picture %s caused Exception %s" %
-                               (filename, e))
-                im = gtk.Label()
-                im.set_text(_("%s" % e))
-            result.append(im)
-        return result
 
     @classmethod
     def get_delimiter(cls, refresh=False):
